+++ app/helpers/holdings_helper.rb
@@ -44,14 +44,12 @@ module HoldingsHelper
   READ_ON_SITE_LOCATION_RCRF = 'UARCRF'.freeze # Research and Collections Resource Facility
   READ_ON_SITE_LOCATION_BPSC = 'UASPCOLL'.freeze # Bruce Peel Special Collections
 
-  def read_on_site_path(item, document_id, document_title)
-    return unless item[:status] == 'READONSITE'
+  def request_rcrf_read_on_site_form?(item)
+    item[:location] == READ_ON_SITE_LOCATION_RCRF && item[:status] == 'READONSITE'
+  end
 
-    if item[:location] == READ_ON_SITE_LOCATION_RCRF
-      new_rcrf_read_on_site_request_path(item_url: catalog_url(document_id), title: document_title)
-    elsif item[:location] == READ_ON_SITE_LOCATION_BPSC
-      new_bpsc_read_on_site_request_path(item_url: catalog_url(document_id), title: document_title)
-    end
+  def request_bpsc_read_on_site_form?(item)
+    item[:location] == READ_ON_SITE_LOCATION_BPSC && item[:status] == 'READONSITE'
   end
 
   def request_microform?(item)